requirements-dev.txt
requests<3,>=2.31.0
requirements-parser<2,>=0.5.0
rich<14,>=13.4.2
pre-commit==3.3.3
